finalising chemdner branch
